Initial TypeScript setup
