+++ .gitignore
@@ -95,3 +95,5 @@ public/
 .dynamodb/
 
 # End of https://www.gitignore.io/api/node
+
+dist/
